Lots of fixes to env pre-docker. Refactored analysis
