+++ Makefile
@@ -50,17 +50,33 @@ data/processed/character_features_deploy.csv : \
 	    -i data/processed/clean_characters.csv \
 	    -o data/processed/character_features.csv -v
 
-# Machine learning modelling
-results/figures/optimized_model.png \
-results/figures/model_comparison.png \
-results/tables/optimized_model.pkl \
+# Machine learning model selection
 results/tables/model_comparison.pkl \
-results/models/optimized_model.pkl : \
+results/figures/model_comparison.png : \
     data/processed/character_features_train.csv \
-	src/analysis.py
-	    python src/analysis.py \
+	src/model_selection.py
+	    python src/model_selection.py \
 		    -i data/processed/character_features_train.csv \
 			-o results -v
+
+# Machine learning model optimization
+results/tables/optimized_model.pkl \
+results/models/optimized_model.pkl : \
+    results/figures/model_comparison.png \
+	src/model_optimize.py
+	    python src/model_optimize.py \
+		    -i data/processed/character_features_train.csv \
+			-o results -v
+
+# Machine learning model optimization no neutrals
+results/figures/polarized_optimized_model.png \
+results/tables/polarized_optimized_model.pkl \
+results/models/polarized_optimized_model.pkl : \
+    results/models/optimized_model.pkl \
+	src/model_optimize.py
+	    python src/model_optimize.py \
+		    -i data/processed/character_features_polar_train.csv \
+			-o results -f polarized_ -v
 
 # Feature importance analysis
 results/figures/importance_of_appearances.png \
@@ -76,9 +92,8 @@ results/tables/importance_of_hair.pkl \
 results/tables/importance_of_id.pkl \
 results/tables/importance_of_publisher.pkl \
 results/tables/importance_of_sex.pkl \
-results/tables/importance_of_year.pkl \
 results/tables/importance_of_year.pkl : \
-    results/tables/optimized_model.pkl \
+    results/tables/polarized_optimized_model.pkl \
 	src/analysis_feature.py
 		python src/analysis_feature.py \
 	    	-i results/tables/optimized_model.pkl \
@@ -91,16 +106,12 @@ report/summary_report.md : \
 	results/figures/appearances_by_alignment.png \
 	results/tables/dataset_overview.pkl \
 	results/tables/feature_overview.pkl \
-	results/figures/optimized_model.png \
     results/figures/model_comparison.png \
-    results/tables/optimized_model.pkl \
-    results/tables/model_comparison.pkl \
 	results/figures/importance_of_eye.png \
 	results/figures/importance_of_hair.png \
 	results/figures/importance_of_id.png \
 	results/figures/importance_of_publisher.png \
-	results/figures/importance_of_sex.png \
-	results/models/optimized_model.pkl
+	results/figures/importance_of_sex.png
 	    jupyter nbconvert --to html report/summary_report.ipynb --no-input
 
 clean :
